DSC Python Study Group  Week 3 Assessment
Stadium Seating

There are three seating categories at a stadium. Class A seats cost $20, Class B seats cost $15, and Class C seats cost $10. Write a function named earnings that asks how many tickets for each class of seats were sold, then returns the amount of income generated from ticket sales. Take care of any possible errors/exceptions that may arise depending on user input (Display an error message).


Prime Numbers
A prime number is a number that is only evenly divisible by itself and 1. For example, the number 5 is prime because it can only be evenly divided by 1 and 5. The number 6, however, is not prime because it can be divided evenly by 1, 2, 3, and 6. Write a Boolean function named is_prime which takes an integer as an argument and returns true if the argument is a prime number, or false otherwise. Use the function in a program that prompts the user to enter a number then displays a message indicating whether the number is prime.  Take care of any possible errors/exceptions that may arise depending on user input (Display an error message).


Prime Factors
Based on your is_prime function in question 2, write another program that determines all of the prime factors of an inputted number and stores them in a .txt file named Results. The program should have a loop that calls the is_prime function.  Take care of any possible errors/exceptions that may arise depending on user input (Display an error message).


File Head Display
Write a program that asks the user for the name of a file. The program should display only the first five lines of the file’s contents. If the file contains less than five lines, it should display the file’s entire contents. Take care of any possible errors/exceptions that may arise depending on the file’s content or the lack thereof (Display an error message).